tests/wpt/web-platform-tests/css/css-scoping/slotted-link.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Scoping: ::slotted :link</title>
<link rel="author" title="Rune Lillesveen" href="mailto:futhark@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-scoping/#slotted-pseudo">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="host"><a id="a" href="#notvisited">This link should be green.</a></div>
<script>
  let root = host.attachShadow({mode:"open"});
  root.innerHTML = `
  <style>
    ::slotted(:link) { color:green }
    ::slotted(:visited) { color:red }
  </style>
  <slot></slot>`;

  test(() => {
    assert_equals(getComputedStyle(a).color, "rgb(0, 128, 0)", "Unvisited link should be green.");
  }, "Check that we match :link and not :visited for slotted anchor.");
</script>
